Shitty workaround for directed graphs with effectively bi-directional edges
This was done to correctly display directed graphs that have effectively bi-directional edges, without that change - only one direction in directional graphs was supported (the other one was simply omitted). Now it displays it correctly (the arrow is on two sides) It is shitty because the edges labels(eg. 1:7 and 7:1) in these cases are superimposed on each other - and it in turn makes them less intelligible.